Gunnar: -embedding for Hofstadter model         -caching environment variables in configure.ac         -small edits for compilation errors
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3267 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/FermionOnTorusWithMagneticTranslations.cc
@@ -1472,7 +1472,7 @@ ComplexVector& FermionOnTorusWithMagneticTranslations::CoreC4Rotation (ComplexVe
 #else
 	  for (int k = 0; k < this->NbrFermions; ++k)
 	    for (int l = 0; l < this->NbrFermions; ++l)
-	      DeterminantMatrix[k][l] = PhaseMatrix[TmpInputMonomial[k]][TmpOutputMonomial[l]];
+	      DeterminantMatrix[k][l] = PhaseMatrix[TmpInputMonomial[k]][(long)TmpOutputMonomial[l]];
 	  Tmp += inputState[j] * TmpCoefficient * DeterminantMatrix.Determinant() * sqrt((double) TmpInputSpace->NbrStateInOrbit[j]);
 #endif
 	}
